% Auto-generated: do not edit by hand
\name{tangoParameterSelector}

\alias{tangoParameterSelector}

\title{TangoParameterSelector component}

\description{
@param  {TangoParameterSelector.propTypes} props @returns {JSX.Element} @constructor
}

\usage{
tangoParameterSelector(id=NULL, availableParams=NULL, dictionary=NULL,
selectedParam=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{availableParams}{List with named elements and values of type list with named elements and values of type list with named elements and values of type list of characters. }

\item{dictionary}{List of lists containing elements 'name', 'param'.
those elements have the following types:
  - name (character; optional)
  - param (optional)s. }

\item{selectedParam}{}

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

