% Auto-generated: do not edit by hand
\name{tangoParameterSelector}

\alias{tangoParameterSelector}

\title{TangoParameterSelector component}

\description{
@param  {TangoParameterSelector.propTypes} props @returns {JSX.Element} @constructor
}

\usage{
tangoParameterSelector(id=NULL, availableParams=NULL, selectedParam=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{availableParams}{List with named elements and values of type list with named elements and values of type list of characters. }

\item{selectedParam}{Lists containing elements 'domain', 'family', 'member'.
those elements have the following types:
  - domain (character; optional)
  - family (character; optional)
  - member (character; optional). }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

